+++ src/button/_example/icon-btn/index.js
@@ -1,5 +1,1 @@
-Component({
-  properties: {
-    disabled: Boolean,
-  },
-});
+Component({});