+++ src/navbar/props.ts
@@ -11,11 +11,6 @@ const props: TdNavbarProps = {
     type: Boolean,
     value: true,
   },
-  /** 背景 */
-  background: {
-    type: String,
-    value: '',
-  },
   /** 自定义组件样式 */
   customStyle: {
     type: String,
@@ -35,15 +30,10 @@ const props: TdNavbarProps = {
     type: Boolean,
     value: true,
   },
-  /** 首页图标地址。值为 '' 或者 undefiend 则表示不显示返回图标，值为 'circle' 表示显示默认图标，值为 'slot' 表示使用插槽渲染，值为其他则表示图标地址 */
-  homeIcon: {
-    type: String,
-    value: '',
-  },
-  /** 左侧图标地址，值为 '' 或者 undefiend 则表示不显示返回图标，值为 'arrow-left' 表示显示返回图标，值为 'slot' 表示使用插槽渲染，值为其他则表示图标地址 */
-  leftIcon: {
-    type: String,
-    value: '',
+  /** 是否展示左侧箭头 */
+  leftArrow: {
+    type: Boolean,
+    value: false,
   },
   /** 页面标题 */
   title: {