+++ src/tag/check-tag.ts
@@ -44,7 +44,7 @@ export default class CheckTag extends SuperComponent {
   methods = {
     setClass() {
       const { prefix, classPrefix } = this.data;
-      const { size, shape, closable, disabled, checked, defaultChecked } = this.properties;
+      const { size, shape, variant, closable, disabled, checked, defaultChecked } = this.properties;
       const isChecked = checked || defaultChecked;
       const tagClass = [
         classPrefix,
@@ -52,8 +52,10 @@ export default class CheckTag extends SuperComponent {
         closable ? `${classPrefix}--closable ${prefix}-is-closable` : '',
         disabled ? `${classPrefix}--disabled ${prefix}-is-disabled` : '',
         isChecked ? `${classPrefix}--checked ${prefix}-is-checked` : '',
+        `${classPrefix}--theme-${isChecked ? 'primary' : 'default'}`,
         `${classPrefix}--size-${size || 'medium'}`,
         `${classPrefix}--shape-${shape || 'square'}`,
+        `${classPrefix}--variant-${variant || 'dark'}`,
       ];
       const className = classNames(tagClass);
       this.setData({